move wct tests to test/browser
