+++ .gitignore
@@ -1,4 +1,4 @@
 node_modules
 build
-test/fixtures/build.js
+test/browser/fixtures/build.js
 npm-debug.log
