+++ package.json
@@ -5,12 +5,12 @@
   "main": "build/index.js",
   "scripts": {
     "build": "babel lib -d build",
-    "build-test": "webpack --config test/fixtures/webpack.config.js",
+    "build-test": "webpack --config test/browser/fixtures/webpack.config.js",
     "docs": "rm -rf docs && jsdoc lib -t node_modules/minami -R README-API.md -d docs",
     "prepublish": "npm run build",
     "test": "npm run build-test && npm run test-local",
-    "test-local": "wct --plugin local test/index.html",
-    "test-sauce": "wct --plugin sauce test/index.html"
+    "test-local": "wct --plugin local test/browser/index.html",
+    "test-sauce": "wct --plugin sauce test/browser/index.html"
   },
   "repository": {
     "type": "git",
