+++ test/browser/fixtures/webpack.config.js
