+++ test/browser/index.html
@@ -2,8 +2,8 @@
 <html>
   <head>
     <meta charset="utf-8">
-    <script src="../node_modules/webcomponents.js/webcomponents.js"></script>
-    <script src="../node_modules/web-component-tester/browser.js"></script>
+    <script src="../../node_modules/webcomponents.js/webcomponents.js"></script>
+    <script src="../../node_modules/web-component-tester/browser.js"></script>
     <script src="fixtures/build.js"></script>
   </head>
   <body>
