+++ test/browser/test-component.js
